chore: update example project
