+++ example_project/assets/tsconfig.json
@@ -8,7 +8,7 @@
 
       /* Types for Vite https://vitejs.dev/guide/features#client-types */
       "types": ["vite/client"],
-  
+
       /* Bundler mode */
       "moduleResolution": "bundler",
       "allowImportingTsExtensions": true,
@@ -16,16 +16,16 @@
       "isolatedModules": true,
       "noEmit": true,
       "jsx": "preserve",
-  
+
       /* Linting */
       "strict": true,
       "noUnusedLocals": true,
       "noUnusedParameters": true,
       "noFallthroughCasesInSwitch": true,
-      
+
       "paths": {
         "@/*": ["./*"]
       }
     },
-    "include": ["js/**/*.ts", "js/**/*.js", "js/**/*.tsx", "vue/**/*.vue", "*.config.ts"],
+    "include": ["js/**/*.ts", "js/**/*.js", "js/**/*.tsx", "vue/**/*.vue",  "vue/index.ts", "*.config.ts"],
   }
